VAN-7693 Hide Power settings section on IDEAPAD B50-80 machine

+++ vtr-ui/src/app/components/pages/page-device-settings/children/subpage-device-settings-power/subpage-device-settings-power.component.ts
@@ -64,7 +64,7 @@ export class SubpageDeviceSettingsPowerComponent implements OnInit, OnDestroy {
 	toggleAlwaysOnUsbFlag = false;
 	usbChargingCheckboxFlag = false;
 	powerMode = PowerMode.Sleep;
-	showEasyResumeSection = false;
+	showEasyResumeSection = true;
 	toggleEasyResumeStatus = false;
 	showAirplanePowerModeSection = false;
 	toggleAirplanePowerModeFlag = false;
@@ -875,14 +875,11 @@ export class SubpageDeviceSettingsPowerComponent implements OnInit, OnDestroy {
 	}
 
 	showPowerSettings() {
-
-		if (!(!this.isDesktopMachine || this.showEasyResumeSection)) {
+		if (this.isDesktopMachine || (!this.showEasyResumeSection && !this.alwaysOnUSBStatus.available)) {
 			this.headerMenuItems = this.commonService.removeObjFrom(this.headerMenuItems, 'power');
+			return false;
 		}
-
-		// return !this.isDesktopMachine || this.showEasyResumeSection || this.usbChargingInBatteryModeStatus;
-		return !this.isDesktopMachine || this.showEasyResumeSection;
-		// return !this.isDesktopMachine || this.showEasyResumeSection;
+		return true
 	}
 
 	private getEnergyStarCapability() {
